[wasm-simd] Emit SIMD opcodes as LEB encoded bytes
With https://crrev.com/c/2118476, we can decode multi byte opcodes. So
the module builder should be emitting the correct LEB encoded bytes for
SIMD opcodes.

This also fixes an error discovered by the fuzzer: I added f64x2.add to
the fuzzer, but that's actually a multi-byte opcode, so it resulted in
incorrect bytes generated. This should fix it.

Bug: chromium:1072090
Bug: v8:10258
Change-Id: I0b32ac27aa24d25ee8694dacb12d3d8339d9f839
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/2158005
Reviewed-by: Deepti Gandluri <gdeepti@chromium.org>
Commit-Queue: Zhi An Ng <zhin@chromium.org>
Cr-Commit-Position: refs/heads/master@{#67257}