Reland "[wasm-simd][liftoff][arm][arm64] Implement integer narrowing"
This relands commit 387d85addaa40da353a07149e15078fa575378e1.

This was passing in the simulator because of the typo in
simulator-arm.cc. The instr->Bit(7,6) call is incorrect, it should be
Bit (no s). I don't really know what it resolves to, but the effect is
that we get src_unsigned is false when we expect it to be true.

With the simulator fix, we get the same error as non-simulator runs. The
fix is in liftoff-assembler-arm.h, we need to be passing both source and
dst data types to vmovqn. This was already done correctly in
code-generator-arm, but I was careless in copying the logic...

Bug: v8:9909

Original change's description:
> [wasm-simd][liftoff][arm][arm64] Implement integer narrowing
>
> Bug: v8:9909
> Change-Id: I0664df45fe399bfa018ff8bcacdbdae66944ed29
> Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/2154834
> Reviewed-by: Clemens Backes <clemensb@chromium.org>
> Commit-Queue: Zhi An Ng <zhin@chromium.org>
> Cr-Commit-Position: refs/heads/master@{#67254}

Change-Id: If17f13aae40569174635283abec5ea1358286f55
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/2157799
Reviewed-by: Clemens Backes <clemensb@chromium.org>
Reviewed-by: Jakob Kummerow <jkummerow@chromium.org>
Commit-Queue: Jakob Kummerow <jkummerow@chromium.org>
Auto-Submit: Zhi An Ng <zhin@chromium.org>
Cr-Commit-Position: refs/heads/master@{#67287}