[compiler] Template methods in SimplifiedLowering pt. 3
This is a CL in a string of CLs that template SimplifiedLowering.
It is done with the purpose of decoupling the implementation in the
three phases that SimplifiedLowering has (Propagate, Retype, Lower).

Template VisitBinop (and Binop specific ones like
VisitSpeculativeInt32Binop), for which we need to template:
 * VisitSpeculativeAdditiveOp
 * VisitSpeculativeNumberModulus
 * VisitCheckBounds

Bug: v8:10424
Change-Id: I51764cb2e9ea9c2aa14829888d352e1947be77b5
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/2135637
Reviewed-by: Georg Neis <neis@chromium.org>
Reviewed-by: Nico Hartmann <nicohartmann@chromium.org>
Commit-Queue: Santiago Aboy Solanes <solanes@chromium.org>
Cr-Commit-Position: refs/heads/master@{#67394}