[isolate] Fix CountUsage calls without a current native context
The blink use counter callback requires a current native context to
exist, and will crash if that is not the case. We can fix this V8 by
deferring the count. A cleaner (future) fix would be to either pass
the native context to the callback, or remove the requirement in
blink.

Bug: v8:9496,v8:10460
Change-Id: I8832d02088ba422c3a27638cee4dacbaaf6f39b4
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/2167394
Reviewed-by: Leszek Swirski <leszeks@chromium.org>
Commit-Queue: Jakob Gruber <jgruber@chromium.org>
Cr-Commit-Position: refs/heads/master@{#67421}