[heap] Remove detached contexts heuristic from the memory reducer
When the memory reducer decides whether to do the next GC or not, it
takes into account the number of detached contexts. This may lead to
redundant GCs if the detached contexts are leaking and stay reachable.

With the unified heap all unreachable detached contexts are freed in
one GC, so this heuristic is no longer necessary.

Bug: chromium:1072746
Change-Id: I869fd6ca2008135dfea39e7ceeb01fdc9ead22be
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/2187493
Reviewed-by: Michael Lippautz <mlippautz@chromium.org>
Commit-Queue: Ulan Degenbaev <ulan@chromium.org>
Cr-Commit-Position: refs/heads/master@{#67653}